+++ R/droplets_create.R
@@ -34,7 +34,7 @@
 #' }
 droplets_create <- function(names = NULL,
                            size = getOption("do_size", "s-1vcpu-1gb"),
-                           image = getOption("do_image", "ubuntu-18-04-x64"),
+                           image = getOption("do_image", "ubuntu-22-04-x64"),
                            region = getOption("do_region", "sfo3"),
                            ssh_keys = getOption("do_ssh_keys", NULL),
                            backups = getOption("do_backups", NULL),